requirements.txt
yt-dlp
requests
python-dotenv
azure-storage-file-datalake
aiofiles
aiohttp
Flask[async]
sounddevice
scipy